requirements.txt
portalocker
chardet
PyYAML==6.0
requests==2.31.0
guessit==3.7.1
tmdbv3api==1.9.0
musicbrainzngs==0.7.1
tqdm==4.66.1
psutil==5.9.5
pathlib~=1.0.1
rapidfuzz~=3.14.1
mutagen~=1.47.0
opensubtitlescom~=0.1.5
knowit~=0.5.11
pandas~=2.3.3
tmdbsimple~=2.9.1
plexapi~=4.17.1
ImageHash~=4.3.2
opencv-python~=4.12.0.88
numpy~=2.2.6
pillow~=12.0.0